Closes #81

See merge request jotunnllc/heimdali/heimdali-api!86

+++ src/main/scala/com/heimdali/models/HiveAllocation.scala
@@ -15,7 +15,7 @@ import scala.concurrent.ExecutionContext
 case class HiveAllocation(name: String,
                           location: String,
                           sizeInGB: Int,
-                          consumedInGB: Option[Double],
+                          consumedInGB: Option[BigDecimal],
                           managingGroup: HiveGrant,
                           readonlyGroup: Option[HiveGrant] = None,
                           id: Option[Long] = None,
